Give rounded numbers (to 5) . Added whitespaces in code for easy editing.
Before, the library gave ridiculous numbers like 0.42548952839458 but now it rounds it to 5 floating digits. Such as 15.16173 # 5 floating points.

Also, When defining magnitudes and order lists it was scattered and difficult to edit. I added whitespaces to make it easier to edit.

Since the library is almost about to have a stable release ill stop making careless commits.